+++ vectorbt_runner/backtest_summary.py
@@ -1,0 +1,10 @@
+"""Backtest summary value object."""
+
+from dataclasses import dataclass
+
+
+@dataclass(slots=True)
+class BacktestSummary:
+    total_combinations: int
+    profitable_combinations: int
+    best_pf: float
